fix TS issue for update input
