+++ apps/api/v2/src/ee/event-types/inputs/update-event-type/fields/location.ts
@@ -1,0 +1,30 @@
+import { IsString, IsOptional, IsBoolean, IsNumber } from "class-validator";
+
+export class Location {
+  @IsString()
+  type!: string;
+
+  @IsString()
+  @IsOptional()
+  address?: string;
+
+  @IsString()
+  @IsOptional()
+  link?: string;
+
+  @IsBoolean()
+  @IsOptional()
+  displayLocationPublicly?: boolean;
+
+  @IsString()
+  @IsOptional()
+  hostPhoneNumber?: string;
+
+  @IsNumber()
+  @IsOptional()
+  credentialId?: number;
+
+  @IsString()
+  @IsOptional()
+  teamName?: string;
+}
