+++ apps/api/v2/src/ee/event-types/inputs/update-event-type/update-event-type.input.ts
@@ -1,3 +1,5 @@
+import { BookingField } from "@/ee/event-types/inputs/update-event-type/fields/booking-field";
+import { Location } from "@/ee/event-types/inputs/update-event-type/fields/location";
 import { Type } from "class-transformer";
 import {
   IsString,
@@ -121,10 +123,11 @@ export class UpdateEventTypeInput {
   @IsOptional()
   parentId?: number;
 
-  @ValidateNested()
-  @Type(() => FieldSchema)
   @IsOptional()
-  bookingFields?: FieldSchema[];
+  @IsArray()
+  @ValidateNested({ each: true })
+  @Type(() => BookingField)
+  bookingFields?: BookingField[];
 
   @IsString()
   @IsOptional()
@@ -254,226 +257,4 @@ export class UpdateEventTypeInput {
   @IsInt()
   @IsOptional()
   secondaryEmailId?: number;
-}
-
-class VariantField {
-  @IsString()
-  @IsOptional()
-  name?: string;
-
-  @IsString()
-  @IsOptional()
-  type?: string; // Adjust according to your actual enum for field types
-
-  @IsString()
-  @IsOptional()
-  label?: string;
-
-  @IsString()
-  @IsOptional()
-  labelAsSafeHtml?: string;
-
-  @IsString()
-  @IsOptional()
-  placeholder?: string;
-
-  @IsBoolean()
-  @IsOptional()
-  required?: boolean;
-}
-
-class Variant {
-  @ValidateNested({ each: true })
-  @Type(() => VariantField)
-  fields!: VariantField[];
-}
-
-class VariantsConfig {
-  @ValidateNested()
-  @Type(() => Variant)
-  variants!: Record<string, Variant>;
-}
-
-class Source {
-  @IsString()
-  id!: string;
-
-  @IsString()
-  type!: string;
-
-  @IsString()
-  label!: string;
-
-  @IsString()
-  @IsOptional()
-  editUrl?: string;
-
-  @IsBoolean()
-  @IsOptional()
-  fieldRequired?: boolean;
-}
-
-class View {
-  @IsString()
-  label!: string;
-
-  @IsString()
-  id!: string;
-
-  @IsString()
-  @IsOptional()
-  description?: string;
-}
-
-enum FieldType {
-  Name = "name",
-  Text = "text",
-  Textarea = "textarea",
-  Number = "number",
-  Email = "email",
-  Phone = "phone",
-  Address = "address",
-  MultiEmail = "multiemail",
-  Select = "select",
-  Multiselect = "multiselect",
-  Checkbox = "checkbox",
-  Radio = "radio",
-  RadioInput = "radioInput",
-  Boolean = "boolean",
-}
-
-enum EditableSchema {
-  System = "system",
-  SystemButOptional = "system-but-optional",
-  SystemButHidden = "system-but-hidden",
-  User = "user",
-  UserReadonly = "user-readonly",
-}
-
-class Option {
-  @IsString()
-  label!: string;
-
-  @IsString()
-  value!: string;
-}
-
-class OptionsInput {
-  @IsEnum(FieldType, { each: true })
-  type!: FieldType;
-
-  @IsOptional()
-  @IsBoolean()
-  required?: boolean;
-
-  @IsOptional()
-  @IsString()
-  placeholder?: string;
-}
-
-class BaseFieldSchema {
-  @IsString()
-  name!: string;
-
-  @IsEnum(FieldType)
-  type!: FieldType;
-
-  @IsOptional()
-  @IsString()
-  label?: string;
-
-  @IsOptional()
-  @IsString()
-  labelAsSafeHtml?: string;
-
-  @IsOptional()
-  @IsString()
-  defaultLabel?: string;
-
-  @IsOptional()
-  @IsString()
-  placeholder?: string;
-
-  @IsOptional()
-  @IsString()
-  defaultPlaceholder?: string;
-
-  @IsOptional()
-  @IsBoolean()
-  required?: boolean;
-
-  @IsOptional()
-  @ValidateNested({ each: true })
-  @Type(() => Option)
-  options?: Option[];
-
-  @IsOptional()
-  @IsString()
-  getOptionsAt?: string;
-
-  @IsOptional()
-  @ValidateNested()
-  @Type(() => OptionsInput)
-  optionsInputs?: { [key: string]: OptionsInput };
-}
-
-class FieldSchema extends BaseFieldSchema {
-  @IsOptional()
-  @IsString()
-  variant?: string;
-
-  @IsOptional()
-  @ValidateNested()
-  @Type(() => VariantsConfig)
-  variantsConfig?: VariantsConfig;
-
-  @IsOptional()
-  @IsArray()
-  views?: View[];
-
-  @IsOptional()
-  @IsBoolean()
-  hideWhenJustOneOption?: boolean;
-
-  @IsOptional()
-  @IsBoolean()
-  hidden?: boolean;
-
-  @IsEnum(EditableSchema)
-  @IsOptional()
-  editable?: EditableSchema = EditableSchema.User;
-
-  @IsOptional()
-  @ValidateNested({ each: true })
-  @Type(() => Option)
-  sources?: Source[];
-}
-
-class Location {
-  @IsString()
-  type!: string;
-
-  @IsString()
-  @IsOptional()
-  address?: string;
-
-  @IsString()
-  @IsOptional()
-  link?: string;
-
-  @IsBoolean()
-  @IsOptional()
-  displayLocationPublicly?: boolean;
-
-  @IsString()
-  @IsOptional()
-  hostPhoneNumber?: string;
-
-  @IsNumber()
-  @IsOptional()
-  credentialId?: number;
-
-  @IsString()
-  @IsOptional()
-  teamName?: string;
 }
