+++ packages/platform/libraries/index.ts
@@ -27,6 +27,7 @@ export { SchedulingType, PeriodType } from "@calcom/prisma/enums";
 export type { EventType } from "@calcom/lib/event-types/getEventTypeById";
 export type { EventTypesByViewer } from "@calcom/lib/event-types/getEventTypesByViewer";
 export type { EventTypesPublic } from "@calcom/lib/event-types/getEventTypesPublic";
+export type { UpdateEventTypeReturn } from "@calcom/trpc/server/routers/viewer/eventTypes/update.handler";
 
 export type PublicEventType = Awaited<ReturnType<typeof getPublicEvent>>;
 export { getPublicEvent };
